requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.2
django-import-export==2.7.1
et-xmlfile==1.1.0
filelock==3.6.0
future==0.18.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
virtualenv==20.13.4
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
